Use `unset` in cmake toolchain file. NFC
The `unset` function is the set as calling `set` without a second
argument but less confusing for the reader.

See https://cmake.org/cmake/help/latest/command/set.html#set-normal-variable

Also move the setting if `CMAKE_DL_LIBS` which it is not related to
`CMakeBackwardCompatibilityC.cmake` (which also had a typo in the name).