Remove WIN32 and APPLE references from Emscripten.cmake. NFC (#25855)
Since we are the toolchain file we decide what to set and no other
toolchain file should be setting these. We have tests that confirm that
these are indeed not set.

Also move the setting if `CMAKE_DL_LIBS` which it is not related to
`CMakeBackwardCompatibilityC.cmake` (which also had a typo in the name).