select: enable to handle timeout
This commit enables the select syscall to handle timeout with multiple event
sources when it is called from a worker.

When a thread worker calls __syscall_newselect, it blocks using
emscripten_proxy_sync_with_ctx. When __syscall_newselect is called with zero
timeout, it is unblocked immediately by calling emscripten_proxy_finish
before returning. When it is called with non-zero timeout,
emscripten_proxy_finish is invoked either by the underlying stream
implementation (where an event occurs) or by a setTimeout callback when the
tiemout expires.

In proxying.c, several wrapper functions for proxying-related APIs are added
to allow the JS implementation of newselect to use them.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>