Enable caching of generated JS output
Logical extension of what we were already doing for symbol lists.  This
caches the whole output of the JS compiler and re-uses it when settings
and JS inputs are unchanged.

This improves link time when the cache is hot (i.e. when the same
program has been linked before).

Hello world:

520ms -> 350ms

Hello world + -sINCLUDE_FULL_LIBRARY:

588ms -> 370ms