Allow synchronous proxying of async JS functions
This new mode allows the proxied work itself to be asynchronous (i.e.
return a promise).  The new behaviour is triggered by marking a
functions as both `__proxy: 'sync'` and also `__async`.

We can use this new mode to replace the bespoke proxying code that
exists for `dlsync` as well as the `poll` system call.  For the poll
system call this bespoke code was added in #25523 and #25990, but can
now be completely removed.

In addition to this, because the `poll` syscall is now marked as
`__async` it automatically works under ASYNCIFY/JSPI too.

One downside of this new mode is that the proxied work requires the main
thread event loop to run.  Unlike the synchronous proxied work that can
complete even when we are deeply nested.

Fixes: #25970