Fix test_embind_tsgen_memory64_pthread (#26017)
I noticed this test was failing locally. This test was being skipped in
CI because the `other` runner skips all wasm64 tests and the the
`wasm64` runner it only runs `other.*_wasm64` tests.

The `_emscripten_thread_mailbox_await` function unconditionally depends
on `waitAsyncPolyfilled` to the symbol needs to at least always exist.